Refactor MarkdownParser: extract BlockParser, InlineParser, and DocumentBuilder
